+++ KubeSnapIt.psm1
@@ -3,68 +3,66 @@
 # MARKER: NEW PARAM BLOCK
 
 # Dot Source all functions in all ps1 files located in this module
-# Define the path to the local Private directory and Krew storage directory for KubeSnapIt
-$localPrivateDir = "$PSScriptRoot\Private"  # Local Private directory
-$krewStorageDir = "$HOME/.krew/store/KubeSnapIt"  # Krew storage directory
+# Define the path to the local Private directory and Krew storage directory for KubeTidy
+$localPrivateDir = "$PSScriptRoot/Private"  # Local Private directory
+$krewStorageDir = "$HOME/.krew/store/kubetidy"  # Krew storage directory
+$foundScripts = $false  # Flag to track if any scripts were found and executed
 
-# Check if the local Private directory exists
+# Check if the local Private directory exists and source scripts if available
 if (Test-Path -Path $localPrivateDir) {
     Write-Verbose "Executing scripts from local Private directory."
 
     # Get all .ps1 files in the local Private directory
-    $localScripts = Get-ChildItem -Path "$localPrivateDir/*.ps1" 2>$null
+    $localScripts = Get-ChildItem -Path "$localPrivateDir/*.ps1"
 
     # Execute each .ps1 script found in the local Private directory
     foreach ($script in $localScripts) {
         Write-Verbose "Executing script: $($script.FullName)"
         . $script.FullName  # Call the script
+        $foundScripts = $true
     }
 }
-else {
-    Write-Verbose "Local Private directory not found, checking Krew storage."
 
-    # Check if the KubeSnapIt storage directory exists
-    if (Test-Path -Path $krewStorageDir) {
-        Write-Verbose "Checking for available versions in $krewStorageDir."
+# If no scripts found in local directory, check Krew storage directory
+if (-not $foundScripts -and (Test-Path -Path $krewStorageDir)) {
+    Write-Verbose "Local Private directory empty or missing. Checking Krew storage directory."
 
-        # Get all version directories (assuming they follow a vX.X.X naming pattern)
-        $versionDirs = Get-ChildItem -Path $krewStorageDir -Directory | Where-Object { $_.Name -match '^v\d+\.\d+\.\d+$' }
+    # Get all version directories (assuming they follow a vX.X.X naming pattern)
+    $versionDirs = Get-ChildItem -Path $krewStorageDir -Directory | Where-Object { $_.Name -match '^v\d+\.\d+\.\d+$' }
 
-        # Check if any version directories were found
-        if ($versionDirs) {
-            # Get the latest version directory based on the version number
-            $latestVersionDir = $versionDirs | Sort-Object { [Version]$_.Name.Substring(1) } -Descending | Select-Object -First 1
+    # Check if any version directories were found
+    if ($versionDirs) {
+        # Get the latest version directory based on the version number
+        $latestVersionDir = $versionDirs | Sort-Object { [Version]$_.Name.Substring(1) } -Descending | Select-Object -First 1
 
-            Write-Verbose "Latest version found: $($latestVersionDir.Name)"
+        Write-Verbose "Latest version found: $($latestVersionDir.Name)"
 
-            # Construct the path to the Private directory for the latest version
-            $kubePrivateDir = Join-Path -Path $latestVersionDir.FullName -ChildPath "Private"
+        # Construct the path to the Private directory for the latest version
+        $kubePrivateDir = Join-Path -Path $latestVersionDir.FullName -ChildPath "Private"
 
-            # Check if the Private directory exists in the latest version
-            if (Test-Path -Path $kubePrivateDir) {
-                # Get all .ps1 files in the Private directory
-                $scripts = Get-ChildItem -Path "$kubePrivateDir/*.ps1" 2>$null
+        # Check if the Private directory exists in the latest version
+        if (Test-Path -Path $kubePrivateDir) {
+            # Get all .ps1 files in the Private directory
+            $scripts = Get-ChildItem -Path "$kubePrivateDir/*.ps1"
 
-                # Execute each .ps1 script found
-                foreach ($script in $scripts) {
-                    Write-Verbose "Executing script: $($script.FullName)"
-                    . $script.FullName  # Call the script
-                }
+            # Execute each .ps1 script found
+            foreach ($script in $scripts) {
+                Write-Verbose "Executing script: $($script.FullName)"
+                . $script.FullName  # Call the script
+                $foundScripts = $true
             }
-            else {
-                Write-Error "No Private directory found for the latest version: $($latestVersionDir.Name). Exiting."
-                exit 1
-            }
+        } else {
+            Write-Verbose "No Private directory found for the latest version: $($latestVersionDir.Name)."
         }
-        else {
-            Write-Error "No version directories found in $krewStorageDir. Exiting."
-            exit 1
-        }
+    } else {
+        Write-Verbose "No version directories found in $krewStorageDir."
     }
-    else {
-        Write-Error "Krew storage directory for KubeSnapIt not found. Exiting."
-        exit 1
-    }
+}
+
+# If no scripts were found and sourced, throw an error
+if (-not $foundScripts) {
+    Write-Error "Error: Unable to source any .ps1 files from either the local Private directory or Krew storage directory. Exiting."
+    exit 1
 }
 
 function Invoke-KubeSnapIt {
@@ -83,6 +81,7 @@ function Invoke-KubeSnapIt {
         [switch]$CompareWithCluster,  # Switch for comparing with the cluster
         [switch]$CompareSnapshots,      # Switch for comparing two snapshots
         [switch]$Force,
+        [switch]$UI,
         [Alias("h")] [switch]$Help
     )
     # END PARAM BLOCK
@@ -105,6 +104,10 @@ function Invoke-KubeSnapIt {
         Write-Host "  -Force               Force the action without prompting for confirmation."
         Write-Host "  -Help                Display this help message."
         return
+    }
+
+    if (!$UI) {
+        Show-KubeSnapItBanner
     }
 
     # Check if kubectl is installed for actions that require it